+++ labbook_BE/app/models/Record.py
@@ -2,7 +2,6 @@
 import logging
 import mysql.connector
 
-from datetime import datetime
 from app.models.Constants import Constants
 from app.models.DB import DB
 from app.models.General import *