+++ labbook_BE/default_settings.py
@@ -1,6 +1,6 @@
 # -*- coding:utf-8 -*-
 
-APP_VERSION = '3.5.8'
+APP_VERSION = '3.5.9'
 
 # ######################################
 #  Database parameters